added logo file. adjusted event card sizing - moved navbar file to its own folder
