+++ react-app/src/components/EventCardMain/EventCardMain.css
@@ -4,7 +4,7 @@
   justify-content: center;
   align-items: center;
   flex-wrap: wrap;
-  max-width: 320px;
+  max-width: 305px;
   margin: 40px 15px;
 }
 
@@ -12,8 +12,8 @@
   color: white;
   justify-content: end;
   position:relative;
-  min-width: 320px;
-  height: 420px;
+  min-width: 300px;
+  height: 360px;
   margin: 15px;
   border-radius: 12px;
   box-shadow:
